+++ cmd/ttn-lw-migrate/main.go
@@ -19,6 +19,7 @@ import (
 
 	_ "go.thethings.network/lorawan-stack-migrate/pkg/source/chirpstack" // ChirpStack source
 	_ "go.thethings.network/lorawan-stack-migrate/pkg/source/ttnv2"      // TTNv2 source
+	_ "go.thethings.network/lorawan-stack-migrate/pkg/source/ttnv3"      // TTS source
 
 	"go.thethings.network/lorawan-stack-migrate/cmd"
 )