+++ pkg/source/ttnv3/ttnv3.go
@@ -1,0 +1,14 @@
+package ttnv3
+
+import (
+	"go.thethings.network/lorawan-stack-migrate/pkg/source"
+)
+
+func init() {
+	source.RegisterSource(source.Registration{
+		Name:        "ttnv3",
+		Description: "Migrate from The Things Stack",
+		FlagSet:     flagSet(),
+		Create:      NewSource,
+	})
+}